root/opt/bazarr_rtl/pyproject.toml
[tool.poetry]
name = "bazarr-rtl"
version = "0.1.0"
description = ""
authors = ["Or Levi <orlevi128@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "bazarr_rtl", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.3"
srt = "^3.5.3"
python-bidi = "^0.4.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
